Fixed comparison typo and added roll limit const to gaming plugin
There was a typo in the new coin flip logic where it was
supposed to run a proper simulation if flips were under
the roll limit but instead would only run if they were over.

Also moved hardcoded roll limit to a constant because it
is used in multiple places.