Added fudge approximation and refactored to match change
This adds the ability to approximate fudge rolls greater
than the roll limit in n_rolls. This also majorly refactors
n_rolls because the changes to fudge dice make it possible
to use the approximation calculations in a more generic
way that is applicable to both fudge and non fudge dice.