Fixed erroring on fudge because 'f' is not a number
When the count is less than 100 the roll function will actually
simulate the dice rolls and if the argument was for fudge dice
then n would be 'f' or 'F' rather than a number. This will cause
random.randint to error.

"Um, Big Bird, F is a letter... not a number"
  -Kramit the Frog, Sensimilla Street