Fixed fudge dice issues with simulated rolls
Because 'F' is a letter and not a number if a person were to roll
a fudge die less than the roll limit the call to random.randint
would fail. This has been replaced with a correct checking
of fudge dice and ensuring the correct values are passed.

"Uh, Big Bird, F is not a number... F is a letter"
  --Kramit the Frog, Sensimilla Street