Moved approximation formulas to separate functions
In order to provide better documentation and readability as to
how rolls are approximated the formulas have been moved to
their own functions. This allows each formula to be better
documented in docstrings as well as giving them concise names
so that there's no reliance on just variable names to understand
how the values are being found.